dmd_big_hack: Diablo
The dmd won't display because it's still using some wmv video files which are not supported in standalone.
I've converted them and adapted the vbs, but I'll close this quickly.
This repo is not for assets, but it will still be possible to find this and I think this can make some fans happy.